Renamed Notification to Multiplayer
Same as last commit, also renamed the api folders benji made. this is discussed with Benji
