+++ app/src/main/java/com/example/prikkie/MainActivityNav.java
@@ -20,7 +20,7 @@ public class MainActivityNav extends AppCompatActivity {
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         AppBarConfiguration appBarConfiguration = new AppBarConfiguration.Builder(
-                R.id.navigation_home, R.id.navigation_weekly_planner, R.id.navigation_notifications, R.id.navigation_account)
+                R.id.navigation_home, R.id.navigation_weekly_planner, R.id.navigation_multiplayer, R.id.navigation_account)
                 .build();
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
         NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
