+++ app/src/main/java/com/example/prikkie/ui/multiplayer/MultiplayerViewModel.java
@@ -1,16 +1,16 @@
-package com.example.prikkie.ui.notifications;
+package com.example.prikkie.ui.multiplayer;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class MultiplayerViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public MultiplayerViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("This is the multiplayer fragment");
     }
 
     public LiveData<String> getText() {
