+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -18,15 +18,15 @@
         tools:layout="@layout/fragment_weekly_planner" />
 
     <fragment
-        android:id="@+id/navigation_notifications"
-        android:name="com.example.prikkie.ui.notifications.NotificationsFragment"
-        android:label="@string/title_notifications"
-        tools:layout="@layout/fragment_notifications" />
+        android:id="@+id/navigation_multiplayer"
+        android:name="com.example.prikkie.ui.multiplayer.MultiplayerFragment"
+        android:label="@string/title_multiplayer"
+        tools:layout="@layout/fragment_multiplayer" />
 
     <fragment
         android:id="@+id/navigation_account"
         android:name="com.example.prikkie.ui.account.AccountFragment"
         android:label="@string/title_account"
-        tools:layout="@layout/fragment_notifications" />
+        tools:layout="@layout/fragment_multiplayer" />
 
 </navigation>
